backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.3
anyio==3.7.1
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
biopython==1.84
bitarray==2.8.1
blinker==1.8.2
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cytoolz==0.12.2
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
EditorConfig==0.12.2
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.3
fasteners==0.14.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
future==0.18.2
greenlet==3.1.1
h11==0.14.0
hexbytes==0.3.1
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==2.8
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsbeautifier==1.14.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lru-dict==1.2.0
macaroonbakery==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
multidict==6.0.4
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
parsimonious==0.9.0
pexpect==4.8.0
Pillow==9.0.1
protobuf==4.24.2
ptyprocess==0.7.0
py-algorand-sdk==2.5.0
pycairo==1.20.1
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.19.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-telegram-bot==20.3
pytz==2022.1
pyunormalize==15.0.0
pyxdg==0.27
PyYAML==5.4.1
referencing==0.30.2
regex==2023.8.8
reportlab==3.6.8
requests==2.22.0
rlp==3.0.0
rpds-py==0.10.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.35
systemd-python==234
texttable==1.6.4
tinyman-py-sdk @ git+https://github.com/tinymanorg/tinyman-py-sdk.git@85b1456731577abf2475896039e3c58b6139343a
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.7
wadllib==1.3.6
web3==6.9.0
websocket-client==1.2.3
websockets==11.0.3
Werkzeug==3.0.4
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
